Install /etc/pam.d/gnome-screensaver
Install it at /etc/pam.d/gnome-screensaver.cdc_integration
and only copy it to /etc/pam.d/gnome-screensaver if it DNE.
This avoids the dependency on the gnome-screensaver package, but does not conflict with it.
If it gets copied, print out that smartcard support needs to be re-applied.

Conflicts:
	SPECS/gnome-shell.spec

Conflicts:
	SPECS/gnome-shell.spec